Version 1.0 Minor code corrections

+++ chords_progressions_transformer_aux.py
@@ -97,7 +97,7 @@ select_model_to_load = "187M-2048E-4L-16H-FP32-Fast-Small" # @param ["93M-1024E-
 
 #@markdown Model precision option
 
-model_precision = "float16" # @param ["bfloat16", "float16", "float32"]
+model_precision = "bfloat16" # @param ["bfloat16", "float16", "float32"]
 
 #@markdown bfloat16 == Half precision/faster speed (if supported, otherwise the model will default to float16)
